+++ packages/components/form-field/src/FormField.tsx
@@ -8,7 +8,7 @@ import { FormFieldProvider } from './FormFieldProvider'
 
 export interface FormFieldProps
   extends ComponentPropsWithoutRef<'div'>,
-    Pick<FormFieldContextState, 'name' | 'isRequired' | 'isInvalid'> {
+    Pick<FormFieldContextState, 'name' | 'state' | 'isRequired'> {
   /**
    * Change the component to the HTML tag or custom component of the only child. This will merge the original component props with the props of the supplied element/component and change the underlying DOM node.
    */
@@ -16,16 +16,16 @@ export interface FormFieldProps
 }
 
 export const FormField = forwardRef<HTMLDivElement, FormFieldProps>(
-  ({ className, name, isInvalid = false, isRequired = false, asChild = false, ...others }, ref) => {
+  ({ className, name, state, isRequired = false, asChild = false, ...others }, ref) => {
     const id = useId()
     const Component = asChild ? Slot : 'div'
 
     return (
-      <FormFieldProvider id={id} name={name} isRequired={isRequired} isInvalid={isInvalid}>
+      <FormFieldProvider id={id} name={name} isRequired={isRequired} state={state}>
         <Component
           ref={ref}
           data-spark-component="form-field"
-          className={cx(className, 'flex flex-col gap-xl')}
+          className={cx(className, 'flex flex-col gap-sm')}
           {...others}
         />
       </FormFieldProvider>