+++ packages/components/tabs/src/TabsList.styles.ts
@@ -6,15 +6,15 @@ export const listStyles = cva([
   'flex w-full',
   'spark-orientation-horizontal:flex-row',
   'spark-orientation-vertical:flex-col',
-  'u-no-scrollbar spark-orientation-vertical:overflow-x-hidden overflow-y-hidden',
-  'after:border-outline after:flex after:shrink after:grow',
+  'overflow-y-hidden u-no-scrollbar spark-orientation-vertical:overflow-x-hidden',
+  'after:flex after:shrink after:grow after:border-outline',
   'spark-orientation-horizontal:after:border-b-sm',
   'spark-orientation-vertical:after:border-r-sm',
 ])
 
 export const navigationArrowStyles = cva([
   '!h-auto flex-none',
-  'border-outline border-b-sm',
+  'border-b-sm border-outline',
   'outline-none',
-  'focus-visible:ring-outline-high focus-visible:bg-surface-hovered ring-inset focus-visible:border-none focus-visible:ring-2',
+  'ring-inset focus-visible:border-none focus-visible:bg-surface-hovered focus-visible:ring-2 focus-visible:ring-outline-high',
 ])