+++ packages/components/icon/CHANGELOG.md
@@ -3,6 +3,10 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [1.7.5](https://github.com/adevinta/spark/compare/@spark-ui/icon@1.7.4...@spark-ui/icon@1.7.5) (2023-06-21)
+
+**Note:** Version bump only for package @spark-ui/icon
+
 ## [1.7.4](https://github.com/adevinta/spark/compare/@spark-ui/icon@1.7.3...@spark-ui/icon@1.7.4) (2023-06-01)
 
 **Note:** Version bump only for package @spark-ui/icon