+++ packages/components/tabs/CHANGELOG.md
@@ -3,6 +3,14 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [1.2.10](https://github.com/adevinta/spark/compare/@spark-ui/tabs@1.2.9...@spark-ui/tabs@1.2.10) (2023-06-21)
+
+**Note:** Version bump only for package @spark-ui/tabs
+
+## [1.2.9](https://github.com/adevinta/spark/compare/@spark-ui/tabs@1.2.8...@spark-ui/tabs@1.2.9) (2023-06-20)
+
+**Note:** Version bump only for package @spark-ui/tabs
+
 ## [1.2.8](https://github.com/adevinta/spark/compare/@spark-ui/tabs@1.2.7...@spark-ui/tabs@1.2.8) (2023-06-19)
 
 ### Bug Fixes